+++ src/modules/attendance/routes/AttendRoutes.jsx
@@ -1,6 +1,5 @@
 import AttendancePage from "../page/AttendancePage";
 import StAttendancePage from "../page/StAttendancePage";
-import StQrPage from "../page/StQrPage";
 import AuthRoute from "../../registration/middleware/AuthRoute"; // Assuming AuthRoute is your auth protection component
 
 export default function AttendRoutes() {
@@ -11,10 +10,6 @@ export default function AttendRoutes() {
         {
           path: ":sectionId/", // Route for student attendance page
           element: <AuthRoute allowed_roles={["Student"]}><StAttendancePage /></AuthRoute>, // Wrap with AuthRoute for authentication
-        },
-        {
-          path: ":sectionId/StQr", // Route for student QR page
-          element: <AuthRoute allowed_roles={["Student"]}><StQrPage /></AuthRoute>,  // Wrap with AuthRoute for authentication
         },
       ],
     },
